[feature] Setup done
In theory setup should be done with little to no changes
anymore. Just need to iron out the generator to create a
propagation strategy. The current idea is for me to create
a file system like '1-schema_type-schema_name.sql' for
each thing that needs to be propagated. Then we can just
import all files in bulk, and run them in order.

I want each file to be an isolated piece, such as a single
table, single procedure. This way it can be a versioned
controlled generations in any repo this gets imported into.